+++ src/component/lecture-info/lecture-info.component.jsx
@@ -6,14 +6,16 @@ import { StyledLectureInfo } from './lecture-info.style';
 const LectureInfo = ({ title, id, description, imagePath, youtubeVideoId, youtubePlaylistId }) => {
   const getImagePath = () =>
     imagePath ? imagePath : `http://img.youtube.com/vi/${youtubeVideoId}/0.jpg`;
-  const getToutubePath = () =>
-    `https://www.youtube.com/watch?v=${youtubeVideoId}&list=${youtubePlaylistId}&t=0s`;
+  const getYoutubePath = () => `/video/${youtubeVideoId}`;
+
+  // TODO: use this later on when linking to youtube
+  // `https://www.youtube.com/watch?v=${youtubeVideoId}&list=${youtubePlaylistId}&t=0s`;
 
   const headerTitlePropList = {
     text: title,
     tag: 'a',
     fontSize: '13px',
-    link: getToutubePath()
+    link: getYoutubePath()
   };
 
   return (