+++ src/component/web-info/web-info.component.jsx
@@ -1,9 +1,10 @@
-import React from 'react'
-import CoursePanel from '../course-panel'
+import React from 'react';
+import Dashboard from '../dashboard';
 import Home from '../home';
 import Course from '../course';
 import { StyledWebInfo, StyledRouter } from './web-info.style';
 import Lecture from '../lecture';
+import PlayVideo from '../play-video';
 
 const WebInfo = () => {
   return (
@@ -11,13 +12,14 @@ const WebInfo = () => {
       <StyledWebInfo>
         <StyledRouter>
           <Home path="/" />
-          <CoursePanel path="dashboard" />
+          <Dashboard path="dashboard" />
           <Course path="/course/:courseId" />
           <Lecture path="/lecture/:lectureId" />
+          <PlayVideo path="/video/:youtubeVideoId" />
         </StyledRouter>
       </StyledWebInfo>
     </>
-  )
-}
+  );
+};
 
-export default WebInfo
+export default WebInfo;